FIX, TEST: Get tests running, remove duplicates, and lint
