--- test/client/user-home.spec.js
@@ -1,18 +1,0 @@
-/* global describe beforeEach it */
-
-import { expect } from 'chai';
-import React from 'react';
-import { shallow } from 'enzyme';
-import { UserHome } from '../../client/components/user-home';
-
-describe('UserHome', () => {
-  let userHome;
-
-  beforeEach(() => {
-    userHome = shallow(<UserHome email={'cody@email.com'} />);
-  });
-
-  it('renders the email in an h3', () => {
-    expect(userHome.find('h3').text()).to.be.equal('Welcome, cody@email.com');
-  });
-});
