+++ test/components/user-home.spec.js
@@ -1,18 +1,19 @@
+/* eslint-disable react/jsx-filename-extension */
 /* global describe beforeEach it */
 
-import {expect} from 'chai'
-import React from 'react'
-import {shallow} from 'enzyme'
-import {UserHome} from './user-home'
+import { expect } from 'chai';
+import { shallow } from 'enzyme';
+import React from 'react';
+import { UserHome } from '../../client/components/user-home';
 
 describe('UserHome', () => {
-  let userHome
+  let userHome;
 
   beforeEach(() => {
-    userHome = shallow(<UserHome email={'cody@email.com'} />)
-  })
+    userHome = shallow(<UserHome email={'cody@email.com'} />);
+  });
 
   it('renders the email in an h3', () => {
-    expect(userHome.find('h3').text()).to.be.equal('Welcome, cody@email.com')
-  })
-})
+    expect(userHome.find('h3').text()).to.be.equal('Welcome, cody@email.com');
+  });
+});
