+++ test/store/user.spec.js
@@ -1,52 +1,52 @@
 /* global describe beforeEach afterEach it */
 
-import {expect} from 'chai'
-import {me, logout} from './user'
-import axios from 'axios'
-import MockAdapter from 'axios-mock-adapter'
-import configureMockStore from 'redux-mock-store'
-import thunkMiddleware from 'redux-thunk'
-import history from '../history'
+import { expect } from 'chai';
+import axios from 'axios';
+import MockAdapter from 'axios-mock-adapter';
+import configureMockStore from 'redux-mock-store';
+import thunkMiddleware from 'redux-thunk';
+import history from '../../client/history';
+import { me, logout } from '../../client/store/userReducer';
 
-const mockAxios = new MockAdapter(axios)
-const middlewares = [thunkMiddleware]
-const mockStore = configureMockStore(middlewares)
+const mockAxios = new MockAdapter(axios);
+const middlewares = [thunkMiddleware];
+const mockStore = configureMockStore(middlewares);
 
 describe('thunk creators', () => {
-  let store
+  let store;
 
-  const initialState = {user: {}}
+  const initialState = { user: {} };
 
   beforeEach(() => {
-    store = mockStore(initialState)
-  })
+    store = mockStore(initialState);
+  });
 
   afterEach(() => {
-    store.clearActions()
-  })
+    store.clearActions();
+  });
 
   describe('me', () => {
     it('eventually dispatches the GET USER action', () => {
-      const fakeUser = {email: 'Cody'}
-      mockAxios.onGet('/auth/me').replyOnce(200, fakeUser)
+      const fakeUser = { email: 'Cody' };
+      mockAxios.onGet('/auth/me').replyOnce(200, fakeUser);
       return store.dispatch(me())
         .then(() => {
-          const actions = store.getActions()
-          expect(actions[0].type).to.be.equal('GET_USER')
-          expect(actions[0].user).to.be.deep.equal(fakeUser)
-        })
-    })
-  })
+          const actions = store.getActions();
+          expect(actions[0].type).to.be.equal('GET_USER');
+          expect(actions[0].user).to.be.deep.equal(fakeUser);
+        });
+    });
+  });
 
   describe('logout', () => {
     it('logout: eventually dispatches the REMOVE_USER action', () => {
-      mockAxios.onPost('/auth/logout').replyOnce(204)
+      mockAxios.onPost('/auth/logout').replyOnce(204);
       return store.dispatch(logout())
         .then(() => {
-          const actions = store.getActions()
-          expect(actions[0].type).to.be.equal('REMOVE_USER')
-          expect(history.location.pathname).to.be.equal('/login')
-        })
-    })
-  })
-})
+          const actions = store.getActions();
+          expect(actions[0].type).to.be.equal('REMOVE_USER');
+          expect(history.location.pathname).to.be.equal('/login');
+        });
+    });
+  });
+});
